% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_icg.R
\name{compute_icg}
\alias{compute_icg}
\title{Computa el ICG (\emph{Compute ICG})}
\usage{
compute_icg(data, wave = NULL, segment_by = NULL)
}
\arguments{
\item{data}{un tibble guardado como objeto en el enviroment luego de consultar \code{\link{get_icg_raw}}
(\emph{A tibble saved as an object in the enviroment after querying \code{\link{get_icg_raw}}}).}

\item{wave}{es el número de ola para la que se quiere computar el ICG
(\emph{number of wave to which compute the ICG}).}

\item{segment_by}{parámetro para computar el ICG segmentado por distintas variables
(\emph{parameter to compute ICG segmented by different variables}).}
}
\value{
devuelve un tibble con \code{class "tbl_df","tbl", "data.frame"} con el valor del ICG y de cada una de sus dimensiones por una o más olas  y segmentado por las variables deseadas
(\emph{returns a tibble with \code{class "tbl_df", "tbl", "data.frame"} with the value of the ICG and the dimensions for one or more waves and segmented by desired variables}).
}
\description{
Función para computar resultados del ICG
(\emph{Function to compute ICG results})
}
\seealso{
\code{\link{get_icg_raw}}
}
