% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icg_raw.R
\name{get_icg_raw}
\alias{get_icg_raw}
\title{Descarga base de microdatos del ICG
(\emph{(Universidad Torcuato Di Tella)})}
\usage{
get_icg_raw()
}
\value{
Devuelve un objeto "tbl_df" "tbl" "data.frame" con los microdatos del ICG....
}
\description{
Función que ...
}
\details{
Respecto el origen de los datos se puede consultar la documentación en
\href{https://www.utdt.edu/ver_contenido.php?id_contenido=1351&id_item_menu=2970}{\code{https://www.utdt.edu/ver_contenido.php?id_contenido=1351&id_item_menu=2970}}.
}
