% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opinAr.R
\docType{package}
\name{opinAr}
\alias{opinAr}
\title{\code{opinAr} package}
\description{
Caja de Herramientas para el procesamiento de la base del ICG
See the README on
\href{https://github.com/PoliticaArgentina/opinAr/blob/master/README.md}{Github}
}
