% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_codebook.R
\name{show_codebook}
\alias{show_codebook}
\title{Muestra libro de códigos
(\emph{(Show codebook)})}
\usage{
show_codebook(data = NULL, viewer = FALSE)
}
\arguments{
\item{data}{un tibble guardado como objeto en el enviroment luego de consultar \code{\link{get_icg_raw}}
(\emph{A tibble saved as an object in the enviroment after querying \code{\link{get_icg_raw}}}).}

\item{viewer}{para elegir si la tabla se ve en el visor o se imprime el tibble para ser utilizado
(\emph{choose to view the table in the viewer or print the tibble to be used}).}
}
\value{
Devuelve un objeto "tbl_df" "tbl" "data.frame" con el libro de códigos del ICG o una "sjTable" con el nombre de las variables, las etiquetas, los valores y las etiquetas de los valores
(\emph{object "tbl_df" "tbl" "data.frame" with the ICG codebook or a "sjTable" with the variables' names, labels, values and value labels})
}
\description{
Función para explorar la estructura de datos del ICG
(\emph{Function to explore the ICG data structure})
}
\details{
Función para obtener la tabla con el libro de códigos
(\emph{Function to show the codebook table})
}
